+++ NEWS.md
@@ -1,3 +1,23 @@
+# Version 0.15.0 [2023-10-28]
+
+## New Features
+
+ * Add `handler_ntfy()` for reporting on progress via the Ntfy.sh
+   Messaging Service using the **ntfy** package.
+
+
+## Bug Fixes
+
+ * `withProgressShiny()` could produce an `if (config$max_steps ==
+   0) : ... argument is of length zero` error.
+
+ * `handlers(new_handlers)` would return `NULL`, instead of `list()`,
+   if there were no prior handlers set.
+
+ * `handler_cli(..., format_done = "...", clear = TRUE)` would not
+   render the `format_done` message, if set.
+
+
 # Version 0.14.0 [2023-08-10]
 
 ## New Features
@@ -609,7 +629,7 @@
  * TESTS: Increased package test coverage of progression handlers by
    running all code except the last step that calls the backend, which
    may not be installed or supported on the current platform,
-   e.g. **tcltk**, **beepr**, notifier.
+   e.g. **tcltk**, **beepr**, and **notifier**.
 
 ## Bug Fixes